requirements.txt
newsapi-python
openai
gradio